Consolidate test files - Phase 1 implementation
- Consolidated entity tracking flow tests from 3 files into 1
  - Merged entity_tracking_flow_test_basic.py and test_entity_tracking_flow_service.py into entity_tracking_flow_test.py
  - Preserved all unique tests including error handling and state logging
  - Used parameterized tests to reduce duplication
  - Removed ~400 lines of duplicate test code

- Merged CRUD extended tests using parameterization
  - test_feed_processing_log_extended.py merged into test_feed_processing_log.py
  - test_rss_feed_extended.py merged into test_rss_feed.py
  - Used pytest.mark.parametrize for test variations
  - Maintained 100% test coverage

All tests passing with no functionality lost.
