+++ test-duplication-analysis.md
@@ -1,0 +1,120 @@
+# Test Duplication Analysis
+
+## Overview
+
+This analysis identifies significant test duplication and redundancy in the Local Newsifier test suite. The test suite contains 105 test files with substantial overlap, particularly in areas of CRUD operations, service layer testing, and workflow testing.
+
+## Key Findings
+
+### 1. Duplicate Test Files Pattern
+
+Multiple test files exist for the same components with suffixes like:
+- `_extended` - Contains additional tests that could be merged
+- `_impl` - Implementation-specific tests that duplicate base tests
+- `_comprehensive` - Overlapping with base tests
+- `_basic` - Simplified versions of main test files
+
+#### Examples:
+- `test_feed_processing_log.py` + `test_feed_processing_log_extended.py`
+- `test_apify_service.py` + `test_apify_service_extended.py` + `test_apify_service_impl.py` + `test_apify_service_schedules.py`
+- `entity_tracking_flow_test.py` + `entity_tracking_flow_test_basic.py` + `test_entity_tracking_flow_service.py`
+
+### 2. Entity Tracking Flow Tests (714 lines across 3 files)
+
+**70% overlapping coverage** identified:
+
+#### Duplicate Test Scenarios:
+1. **Flow Initialization** - Tested in all 3 files
+2. **Process Method** - Tested in 2 files
+3. **Process Article** - Tested in 2 files
+4. **Process New Articles** - Tested in 2 files
+5. **Entity Dashboard** - Tested in 2 files
+6. **Entity Relationships** - Tested in 2 files
+
+#### Unique Tests (worth preserving):
+- Error handling tests in `test_entity_tracking_flow_service.py`
+- State logging tests in `test_entity_tracking_flow_service.py`
+
+### 3. Service Layer Over-Mocking
+
+Service tests contain excessive mocking (120+ mocks in some files):
+- `test_entity_service.py` - 120 mock calls
+- `test_entity_service_extended.py` - 120 mock calls
+- Both files test similar scenarios with slightly different data
+
+### 4. CRUD Test Duplication
+
+Pattern of basic + extended tests for CRUD operations:
+- Basic tests cover standard CRUD operations
+- Extended tests add minor variations that could be parameterized
+- Example: `test_feed_processing_log.py` tests basic operations while `_extended` adds custom parameters
+
+### 5. CLI Test Redundancy
+
+Multiple CLI test files with overlapping coverage:
+- `test_db.py` + `test_db_comprehensive.py`
+- `test_feeds.py` + `test_feeds_container.py`
+- Injectable provider tests spread across multiple files
+
+### 6. Tool Implementation Tests
+
+Duplicate patterns for tools:
+- `test_trend_analyzer.py` + `test_trend_analyzer_impl.py`
+- `test_opinion_visualizer.py` + `test_opinion_visualizer_impl.py`
+- `test_web_scraper.py` + `test_web_scraper_impl.py`
+
+## Quantitative Analysis
+
+### Test File Distribution:
+- **Total test files**: 105
+- **Files with duplication indicators**: 26 (25%)
+- **Files using @injectable**: 6 (being phased out)
+- **Heavy mocking (>40 mocks)**: 12 files
+
+### Coverage Overlap Estimates:
+- **Entity tracking flows**: 70% overlap
+- **Service tests**: 60% overlap
+- **CRUD tests**: 40% overlap
+- **CLI tests**: 50% overlap
+
+## Impact
+
+1. **Maintenance Burden**: Changes require updating multiple test files
+2. **CI/CD Performance**: Redundant tests increase build time
+3. **Code Clarity**: Difficult to understand which tests are authoritative
+4. **False Confidence**: Multiple tests for same functionality create illusion of thorough testing
+
+## Recommendations
+
+### 1. Consolidate Duplicate Files
+Merge `_extended`, `_impl`, `_basic` variations into main test files using:
+- Parameterized tests for variations
+- Test classes to group related tests
+- Descriptive test names to indicate coverage
+
+### 2. Reduce Mock Complexity
+- Use fixtures for common mock setups
+- Consider integration tests over heavily mocked unit tests
+- Mock at service boundaries, not internal components
+
+### 3. Standardize Test Organization
+- One test file per component
+- Clear separation between unit and integration tests
+- Consistent naming without suffixes
+
+### 4. Remove Injectable Tests
+As per migration plan, remove @injectable decorator tests and replace with:
+- Native FastAPI dependency injection tests
+- HTTP client tests for CLI components
+
+### 5. Implement Test Coverage Tools
+- Use coverage.py to identify actual coverage
+- Remove tests that don't increase coverage
+- Focus on testing behavior, not implementation
+
+## Estimated Reduction
+
+If implemented, these changes could reduce:
+- **Test file count**: From 105 to ~75 files (28% reduction)
+- **Test execution time**: Estimated 30-40% faster
+- **Code maintenance**: ~2000 lines of redundant test code removed
